establishes successful login
using the instantiator crate to test, the xmlrpc message now
goes through successfully and can log in a user on a live simulator!!
Cannot log out yet,which I will deal with in the session crate. but
for now, the login is successful! My many years of torment are OVER!!!