gets udp io working
headers are getting formatted correctly, and the mailbox is able to send
and receieve data. Circuitcode packet is readable by the server, and the
ack is getting received by the client. Now the next step is to clean up
the header builder file, read the ack, and set up sending a logout
message!!