Adds packet queue
begins handling acks, and not spamming the login server unnecessarily.
Jesus christ this code is a disaster, but I'm amazed that it works
we're doing it, babey
now that acks are working, I have to investigate what the deal is with
the id of the circuitcode being 1

next stop is generalizing this for packets that require acks, sending
the CompleteAgentMovement packet, and sending the AgentUpdate packet.
Once that is done our login will be fully up to spec