fixes error handling, rectifies hubris
I deleted the xml-rpc-rs repo from the benthic project thinking it was
an embarrassing accident that I published a crate for and everything,
despite not actually needing it at all
apparently I did need it, I did do the right thing by making the fork,
it was useful, and by deleting it you break the entire system when it
tries to parse an i4. I am stronger now for having done this. Despite
having wasted many hours of my life on it.

this commit also majorly improves error handling to make writing the ui
in godot less bad

fix