creates chat crate
I'm making this chat crate to help debug things
it isn't great, since you can't type and see the output at the same
time, but it's helping me figure out why godot isn't working (it's
because of my lack of understanding of actix lifetimes. I thought
spawning one from a blocking runtime would spawn it off and let it do
its own thing, but that is actually not true)
also implements acks for packets that need them. I don't think that the
ack structure is right, but it's working