gets acks working properly!
However, the timing is a little off.
I'm waiting on the send, instead of spinning it out into its own thread,
so it blocks the ack verify process just long enough for it to be a
problem. It's time to go through and refactor things before moving onto
the next steps.