moves to bevy_egui
This is a much more functional option. Honestly this should be part of
built in bevy. This rules.
Now that I have my login working properly, I can use the login button to
trigger the initializaion of the session normally. The current setup
does NOT work, because the login data is empty since the actix session
starter starts up at startup when that struct is empty