Major refactor, message passing to UDS
rewrote major sections of the codebase to use UDS for message passing.
This allows for the client frontend and the client backend to run
seperately without blocking each other, and make better use of the actix
actor system. Next steps are to figure out the schema for sending data
from the server to the UI, and what data the UI needs.