fixes directory structure
the login crate was a bad idea from the start. it's a message,
obviously. it's sending a message, even if all the logic surrounding it
is bizarre. I removed the crate, along with the totally busted chat
crate. I also moved all of the message types out of the messages/models
dir because what was I doing with all that tbh