Gets Chat Working
chat to and from the UI to server is working.
However, parsing might be a little strange from the ChatFromServer
object.
Check to be extra sure that you're not dropping bytes if it expects to
be length-prefixed. It might, it might not.
also osgrid login does not work!
login.osgrid.org is timing out! I am not sure why!