Handles objects, prepares for meshes.
We are now officially at the point where we are receiving the triangles.
Lots of things need to be cleaned up here. Nothing is
documented, and there are plenty of things that I should really be
re-architecting. But my main task now is to find out how to conver the
llmesh binary to gltf. Once that's done, things are going to get nuts.